+++ src/main/java/com/example/tiggle/repository/university/UniversityRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface UniversityRepository extends JpaRepository<University, Integer> {
+public interface UniversityRepository extends JpaRepository<University, Long> {
 
     List<University> findAllByOrderByNameAsc();
 }
